requirements.txt
PySide6>=6.7,<6.11
requests>=2.31
cryptography>=41
qrcode[pil]>=7.4
Pillow>=10
pywin32>=306
openpyxl>=3.1
